+++ src/api/node/attributes/txkernelprintable.cpp
@@ -1,5 +1,8 @@
 #include "txkernelprintable.h"
 
+/**
+ * @brief TxKernelPrintable::TxKernelPrintable
+ */
 TxKernelPrintable::TxKernelPrintable() :
     m_feeShift(0),
     m_fee(0),
@@ -7,69 +10,118 @@ TxKernelPrintable::TxKernelPrintable() :
 {
 }
 
-// Getter Implementierungen
+/**
+ * @brief TxKernelPrintable::features
+ * @return
+ */
 QString TxKernelPrintable::features() const
 {
     return m_features;
 }
 
-uint8_t TxKernelPrintable::feeShift() const
+/**
+ * @brief TxKernelPrintable::feeShift
+ * @return
+ */
+quint64 TxKernelPrintable::feeShift() const
 {
     return m_feeShift;
 }
 
+/**
+ * @brief TxKernelPrintable::fee
+ * @return
+ */
 quint64 TxKernelPrintable::fee() const
 {
     return m_fee;
 }
 
+/**
+ * @brief TxKernelPrintable::lockHeight
+ * @return
+ */
 quint64 TxKernelPrintable::lockHeight() const
 {
     return m_lockHeight;
 }
 
+/**
+ * @brief TxKernelPrintable::excess
+ * @return
+ */
 QString TxKernelPrintable::excess() const
 {
     return m_excess;
 }
 
+/**
+ * @brief TxKernelPrintable::excessSig
+ * @return
+ */
 QString TxKernelPrintable::excessSig() const
 {
     return m_excessSig;
 }
 
-// Setter Implementierungen
+/**
+ * @brief TxKernelPrintable::setFeatures
+ * @param features
+ */
 void TxKernelPrintable::setFeatures(const QString &features)
 {
     m_features = features;
 }
 
-void TxKernelPrintable::setFeeShift(uint8_t feeShift)
+/**
+ * @brief TxKernelPrintable::setFeeShift
+ * @param feeShift
+ */
+void TxKernelPrintable::setFeeShift(quint64 feeShift)
 {
     m_feeShift = feeShift;
 }
 
+/**
+ * @brief TxKernelPrintable::setFee
+ * @param fee
+ */
 void TxKernelPrintable::setFee(quint64 fee)
 {
     m_fee = fee;
 }
 
+/**
+ * @brief TxKernelPrintable::setLockHeight
+ * @param lockHeight
+ */
 void TxKernelPrintable::setLockHeight(quint64 lockHeight)
 {
     m_lockHeight = lockHeight;
 }
 
+/**
+ * @brief TxKernelPrintable::setExcess
+ * @param excess
+ */
 void TxKernelPrintable::setExcess(const QString &excess)
 {
     m_excess = excess;
 }
 
+/**
+ * @brief TxKernelPrintable::setExcessSig
+ * @param excessSig
+ */
 void TxKernelPrintable::setExcessSig(const QString &excessSig)
 {
     m_excessSig = excessSig;
 }
 
-// JSON parsing
+/**
+ * @brief TxKernelPrintable::fromJson
+ * @param json
+ */
 void TxKernelPrintable::fromJson(const QJsonObject &json)
 {
     m_features = json.value("features").toString();
@@ -80,7 +132,10 @@ void TxKernelPrintable::fromJson(const QJsonObject &json)
     m_excessSig = json.value("excess_sig").toString();
 }
 
-// JSON creation
+/**
+ * @brief TxKernelPrintable::toJson
+ * @return
+ */
 QJsonObject TxKernelPrintable::toJson() const
 {
     QJsonObject json;
